Documentaçao

+++ src/main/java/Supermercado/Program/Services/ProdutoService.java
@@ -2,6 +2,7 @@ package Supermercado.Program.Services;
 
 import Supermercado.Program.DTO.ProdutoDTO;
 import Supermercado.Program.Entities.Produtos;
+import Supermercado.Program.Repository.FornecedorRepository;
 import Supermercado.Program.Repository.ProdutosRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,7 +16,8 @@ public class ProdutoService {
 
     @Autowired
     private ProdutosRepository produtosRepository;
-
+    @Autowired
+    FornecedorRepository fornecedorRepository;
 
     public List<ProdutoDTO> listarTodos() {
         List<Produtos> produtos = produtosRepository.findAll();
@@ -32,9 +34,16 @@ public class ProdutoService {
 
 
     public ProdutoDTO inserir(ProdutoDTO produtoDTO) {
-        Produtos produto = new Produtos(produtoDTO);  // Converte ProdutoDTO para Produto
+        Produtos produto = new Produtos(produtoDTO);
+
+
+        if (produtoDTO.getFornecedorId() != null) {
+            fornecedorRepository.findById(produtoDTO.getFornecedorId())
+                    .ifPresent(produto::setFornecedor);
+        }
+
         Produtos produtoSalvo = produtosRepository.save(produto);
-        return new ProdutoDTO(produtoSalvo);  // Retorna ProdutoDTO após a inserção
+        return new ProdutoDTO(produtoSalvo);
     }
 
 
@@ -55,9 +64,17 @@ public class ProdutoService {
                 });
     }
 
+    public ProdutoDTO buscarPorId(Integer id) {
+        Produtos produtos = produtosRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Produto não encontrado com ID: " + id));
+        return new ProdutoDTO(produtos);
+    }
 
 
     public void deletar(Integer id) {
+        if (!produtosRepository.existsById(id)) {
+            throw new IllegalArgumentException("Produto com ID " + id + " não existe.");
+        }
         produtosRepository.deleteById(id);
     }
 }
